+++ docker-compose.backend.yml
@@ -6,8 +6,10 @@ services:
     restart: always
     ports:
       - '5432:5432'
-    env_file:
-      - .env
+    environment:
+      POSTGRES_PASSWORD: 'password'
+      POSTGRES_DB: 'Drello'
+      PGDATA: '/var/lib/postgresql/data/pgdata'
     container_name: drello_db
     volumes:
       - ./pg/data:/var/lib/postgresql/data
@@ -19,5 +21,12 @@ services:
     restart: always
     ports:
       - '3000:3000'
-    env_file:
-      - .env
+    environment:
+      DB_PASSWORD: 'password'
+      DB_NAME: 'Drello'
+      DB_PORT: '5432'
+      DB_HOST: db
+      DB_USERNAME: 'postgres'
+      PORT: '3000'
+      STATE: 'dev'
+      JWT_SECRET: 'VQZ8cmf8naf_cwd.ukb'
