+++ src/home/pages/HomePage.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts"></script>
+
+<template>
+  <q-page padding>
+    <h1 class="text-dark">HomePage</h1>
+  </q-page>
+</template>
+
+<style lang="scss" scoped></style>
