+++ src/layouts/MainLayout.vue
@@ -1,11 +1,15 @@
 <script setup lang="ts">
-import Drawer from 'src/components/shared/MenuDrawer.vue';
+import { ref } from 'vue';
+
+const drawer = ref(true);
 </script>
 
 <template>
-  <q-layout view="hHr lpR lfr">
+  <q-layout view="hHh lpR fFf">
     <q-header elevated class="bg-primary text-white">
       <q-toolbar>
+        <q-btn dense flat round icon="o_menu" @click="drawer = !drawer" />
+
         <q-toolbar-title>
           <q-avatar>
             <img src="https://cdn.quasar.dev/logo-v2/svg/logo-mono-white.svg" />
@@ -15,7 +19,9 @@ import Drawer from 'src/components/shared/MenuDrawer.vue';
       </q-toolbar>
     </q-header>
 
-    <Drawer />
+    <q-drawer show-if-above v-model="drawer" side="left" bordered>
+      <!-- drawer content -->
+    </q-drawer>
 
     <q-page-container>
       <router-view />
