+++ src/router/auth.routes.ts
@@ -1,3 +1,4 @@
+import useAuth from 'src/auth/composables/useAuth';
 import { RouteRecordRaw } from 'vue-router';
 
 const authRoutes: RouteRecordRaw[] = [
@@ -5,6 +6,7 @@ const authRoutes: RouteRecordRaw[] = [
     path: '/auth',
     name: 'auth-layout',
     component: () => import('layouts/AuthLayout.vue'),
+    beforeEnter: (_, __, next) => (useAuth().isAuthenticated.value ? next({ name: 'home-page' }) : next()),
     children: [
       { path: '', name: 'auth', redirect: { name: 'login' } },
       {
