+++ src/router/routes.ts
@@ -1,3 +1,4 @@
+import useAuth from 'src/auth/composables/useAuth';
 import { RouteRecordRaw } from 'vue-router';
 import authRoutes from './auth.routes';
 
@@ -5,12 +6,13 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/',
     name: 'home-layout',
+    beforeEnter: (_, __, next) => (useAuth().isAuthenticated.value ? next() : next({ name: 'login' })),
     component: () => import('layouts/MainLayout.vue'),
     children: [
       {
         path: '',
         name: 'home-page',
-        component: () => import('pages/IndexPage.vue'),
+        component: () => import('src/home/pages/HomePage.vue'),
       },
     ],
   },
