--- src/shared/pages/IndexPage.vue
@@ -1,18 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-  <q-page padding class="flex flex-center">
-    <q-card class="my-card">
-      <img src="https://cdn.quasar.dev/img/mountains.jpg" />
-      <q-card-section>
-        <div class="text-h6">Our Changing Planet</div>
-        <div class="text-subtitle2">by John Doe</div>
-      </q-card-section>
-      <q-card-section>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit
-      </q-card-section>
-    </q-card>
-  </q-page>
-</template>
-
-<style lang="scss" scoped></style>
